Crate chalki

Structs

Args
This struct represents the various command line options available.

Error
The Error type.

Program
A loaded and parsed program.

Enums

ErrorKind
The kind of an error.

Constants

USAGE

_IMPL_DESERIALIZE_FOR_Args

Traits

ResultExt
Additional methods for Result, for easy interaction with this crate.

Functions

goal
Parse a goal and attempt to solve it, using the specified solver.

help
Print out help for commands in interpreter mode.

load_program
Load the file into a string, and parse it.

main

process
Process a single command. args is a struct containing the command-line arguments, and prog is a parsed .chalk file.

read_program
Read a program from the command-line. Stop reading when EOF is read. If an error occurs while reading, a Result::Err is returned.

readline_loop
Reads input lines from the user. Lines start with the string given by prompt. Each line the user enters is passed to the function f for processing.

run

Type Definitions

Result
Convenient wrapper around std::Result.